Move the remaining enums closer to where they are used
