+++ components/datetime/src/format/datetime.rs
@@ -6,8 +6,8 @@ use crate::fields::{self, Field, FieldLength, FieldSymbol, Second, Week, Year};
 use crate::input::ExtractedInput;
 use crate::pattern::runtime::{PatternBorrowed, PatternMetadata};
 use crate::pattern::PatternItem;
-use crate::provider::date_time::GetNameForDayPeriodError;
-use crate::provider::date_time::{
+use super::GetNameForDayPeriodError;
+use super::{
     GetNameForMonthError, GetNameForWeekdayError, GetSymbolForEraError,
     MonthPlaceholderValue
 };
